+++ src/main/java/pixelitor/tools/brushes/FixedDistanceSpacing.java
@@ -29,7 +29,7 @@ public class FixedDistanceSpacing implements SpacingStrategy {
     }
 
     @Override
-    public double getSpacing(int radius) {
+    public double getSpacing(double radius) {
         return distance;
     }
 }
