+++ src/main/java/pixelitor/tools/brushes/OutlineSquareBrush.java
@@ -31,7 +31,7 @@ import static java.awt.BasicStroke.JOIN_BEVEL;
  * The "Squares" brush
  */
 public class OutlineSquareBrush extends OutlineBrush {
-    public OutlineSquareBrush(int radius, OutlineBrushSettings settings) {
+    public OutlineSquareBrush(double radius, OutlineBrushSettings settings) {
         super(settings, radius, CAP_SQUARE, JOIN_BEVEL);
     }
