+++ src/main/java/pixelitor/tools/brushes/ShapeDabsBrush.java
@@ -32,7 +32,7 @@ import static java.awt.RenderingHints.VALUE_ANTIALIAS_ON;
  * A {@link DabsBrush} where the dabs are filled shapes
  */
 public class ShapeDabsBrush extends DabsBrush {
-    public ShapeDabsBrush(int radius, ShapeDabsBrushSettings settings) {
+    public ShapeDabsBrush(double radius, ShapeDabsBrushSettings settings) {
         super(radius, settings, false);
     }
